moveonly: Move handle_local_add_channel to routing.h
